+++ Common/BusinessModels/LoadingUnit/LoadingUnitSize.cs
@@ -8,13 +8,13 @@ namespace Ferretto.Common.BusinessModels
     {
         #region Properties
 
-        public int Height { get; set; }
+        public double Height { get; set; }
 
-        public int Length { get; set; }
+        public double Length { get; set; }
 
         public int Weight { get; set; }
 
-        public int Width { get; set; }
+        public double Width { get; set; }
 
         #endregion
     }
